requirements.txt
streamlit>=1.28.0
openai>=1.0.0
pandas>=2.0.0
plotly>=5.18.0
folium>=0.15.0
streamlit-folium>=0.15.0
requests>=2.31.0
python-dotenv>=1.0.0
reportlab>=4.0.0
scipy>=1.11.0
numpy>=1.24.0
